Rename files and reorder functions within them
